196 rm shitty dialog
